Rename : api와 연관없는 더미데이터, 인터페이스 정의 파일들 분리
- 더미데이터 mocks 폴더로 옮김
- type interface 정의 types로 옮김
