+++ src/components/RoleBox.tsx
@@ -1,4 +1,4 @@
-import { RoleItemToShow } from "@api/interface";
+import { RoleItemToShow } from "@/types/shared";
 import styled from "styled-components";
 import splitAgeMinMax from "@utills/splitAgeMinMax";
