+++ src/components/organisms/DateSelectorBar.tsx
@@ -4,7 +4,7 @@ import { useSelector } from "react-redux";
 import { RootState } from "@redux/store";
 
 import { setHomeDate, setScheduleDate } from "@redux/dateSlice";
-import { DateDetailedInfo, DateSelctedType } from "@api/dateInteface";
+import { DateDetailedInfo, DateSelctedType } from "@/types/dateInteface";
 import { styled } from "styled-components";
 
 /**
