+++ src/pages/DateSelectedNoticeListForCom.tsx
@@ -5,7 +5,7 @@ import { useSelector } from "react-redux";
 import { RootState } from "@redux/store";
 import { useEffect, useState } from "react";
 import jobPostAPI from "@api/jobPostAPI";
-import { JobPost } from "@api/interface";
+import { JobPost } from "@/types/shared";
 import Loading from "@components/Loading";
 import NotFoundPage from "@pages/Error/NotFound";
 import { sendMessage } from "@api/utils";
