+++ src/pages/ExtraShootManagePage.tsx
@@ -16,7 +16,7 @@ import {
   ShootManageList,
   ShootManageSelectStatus,
   ShootManage,
-} from "@api/interface";
+} from "@/types/shared";
 import { requestDeleteFetch, requestGetFetch } from "@api/utils";
 
 export default function ExtraShootManagePage() {
